server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
ollama==0.1.4
openai-whisper
pyaudio==0.2.13
numpy==1.26.2
